% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/efficiency.R
\name{efficiency}
\alias{efficiency}
\title{efficiency}
\usage{
efficiency(design, v, b, k, ctrl, type = "D")
}
\arguments{
\item{design}{matrix of a design, b rows and k columns}

\item{v}{number of treatments and controls}

\item{b}{number of blocks}

\item{k}{block size}

\item{ctrl}{Number of controls required in each block, 1 for 1 control, 0 for no control}

\item{type}{Type of efficiency}
}
\value{
The relative efficiency of the given design
}
\description{
This script calculates the relative efficiency of design with given optimality criterion.
}
\examples{
design = generate_eyd(6,5,5,1)
efficiency(design,6,5,5,1, type = "D")
}
