% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_eyd.R
\name{generate_eyd}
\alias{generate_eyd}
\title{generate_eyd}
\usage{
generate_eyd(v, b, k, ctrl)
}
\arguments{
\item{v}{number of treatments and controls}

\item{b}{number of blocks}

\item{k}{block size}

\item{ctrl}{number of controls required in each block, 1 for 1 control, 0 for no control}
}
\value{
An Extended Youden Design
}
\description{
This script generates an Extended Youden design with ctrl controls required in each block, default 1
}
\examples{
\donttest{
   generate_eyd(6,20,3,0)
}
}
