% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/informat.R
\name{informat}
\alias{informat}
\title{informat}
\usage{
informat(design, v, b, k, ctrl)
}
\arguments{
\item{design}{matrix of a design, b rows and k columns}

\item{v}{number of treatments and controls}

\item{b}{number of blocks}

\item{k}{block size}

\item{ctrl}{Number of controls required in each block, 1 for 1 control, 0 for no control}
}
\value{
The information matrix of the design
}
\description{
This script calculates the information matrix of the design
}
\examples{
design = generate_eyd(6,5,5,1)
informat(design,6,5,5,1)
}
