% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdr_exist.R
\name{sdr_exist}
\alias{sdr_exist}
\title{sdr_exist}
\usage{
sdr_exist(x, block_by_row = T)
}
\arguments{
\item{x}{matrix}

\item{block_by_row}{indication of orientation of the finding}
}
\value{
TRUE/FALSE if the remaining design has SDR
}
\description{
This script checks if the remaining design has SDR for each block
}
