% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdr_find.R
\name{sdr_find}
\alias{sdr_find}
\title{sdr_find}
\usage{
sdr_find(x, block_by_row = T)
}
\arguments{
\item{x}{matrix}

\item{block_by_row}{indication of orientation of the finding}
}
\value{
A Set of Distinct Representative of the matrix
}
\description{
This script finds a Set of Distinct Representative of a matrix
}
