% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shuffle.R
\name{shuffle}
\alias{shuffle}
\title{shuffle}
\usage{
shuffle(x)
}
\arguments{
\item{x}{matrix of a design, b rows and k columns}
}
\value{
Shuffled matrixx
}
\description{
This script generates a randome shuffle within each row of a matrix
}
\examples{
X = matrix(c(0,1,2,2,0,1), byrow = TRUE,nrow = 2)
shuffle(X)
}
